+++ annonex2embl/IOOps.py
@@ -20,9 +20,9 @@ from Bio import SeqIO
 ###############
 
 __author__ = 'Michael Gruenstaeudl <m.gruenstaeudl@fu-berlin.de>'
-__copyright__ = 'Copyright (C) 2016-2018 Michael Gruenstaeudl'
+__copyright__ = 'Copyright (C) 2016-2019 Michael Gruenstaeudl'
 __info__ = 'annonex2embl'
-__version__ = '2018.03.26.2000'
+__version__ = '2019.05.15.1500'
 
 #############
 # DEBUGGING #
@@ -36,7 +36,7 @@ import pdb
 ###########
 
 class Inp:
-    ''' This class contains functions to conduct miscellaneous input 
+    ''' This class contains functions to conduct miscellaneous input
         operations.
     Args:
         [specific to function]
@@ -55,9 +55,10 @@ class Inp:
         return fn
 
     def repl_fileend(self, fn, new_end):
-        ''' This function replaces the file ending (e.g. ".csv") with a 
+        ''' This function replaces the file ending (e.g. ".csv") with a
         different ending. '''
         return fn[:fn.rfind('.')] + '.' + new_end
+
 
     def parse_csv_file(self, path_to_csv):
         ''' This function parses a csv file. '''
@@ -76,6 +77,8 @@ class Inp:
             aln.read(path_to_nex)
             charsets = aln.charsets
             matrix = aln.matrix
+        except Nexus.NexusError as ne:
+            raise ne
         except:
             raise ME.MyException('Parsing of .nex-file unsuccessful.')
         return (charsets, matrix)
@@ -96,7 +99,7 @@ class Outp:
 
     def write_EntryUpload(self, seq_record, outp_handle, eusubm_bool):
         ''' This function writes a seqRecord in ENA format for a submission
-            via Entry Upload. Upon request (eusubm_bool), it also masks the ID and AC 
+            via Entry Upload. Upon request (eusubm_bool), it also masks the ID and AC
             lines as requested by ENA for submissions.
         Args:
             seq_record (obj)
@@ -113,7 +116,6 @@ class Outp:
         except:
             raise ME.MyException('%s annonex2embl ERROR: Problem with \
             `%s`. Did not write to internal handle.' % ('\n', seq_name))
-
         if eusubm_bool:
             temp_handle_lines = temp_handle.getvalue().splitlines()
             if temp_handle_lines[0].split()[0] == 'ID':
@@ -135,3 +137,12 @@ class Outp:
         temp_handle.close()
 
         # return something?
+
+    def create_manifest_file(self, path_to_outfile, study, name, description = ""):
+        test = ''.join(path_to_outfile.split('.')[:-1]) + '.manifest'
+        manifest = open(test, "w")
+        manifest.write("STUDY\t" + study + "\n")
+        manifest.write("NAME\t" + name + "\n")
+        if(description != ""):
+            manifest.write("Description\t" + description + "\n")
+        manifest.close()